数据库id字段类型改为bigint

+++ blog-comment/src/main/java/com/mrqinzh/comment/listener/CommentReplyListener.java
@@ -14,6 +14,6 @@ public class CommentReplyListener implements RocketMQListener<CommentVo> {
     @Override
     public void onMessage(CommentVo message) {
 //        logger.info("this is CommentReplyListener ...");
-        Integer articleId = message.getArticleId();
+        Long articleId = message.getArticleId();
     }
 }
